backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "nta-backend"
version = "0.1.0"
description = "FastAPI API and Temporal worker for NTA Platform."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "asyncpg>=0.30.0",
  "boto3>=1.35.0",
  "fastapi>=0.115.0",
  "greenlet>=3.1.1",
  "httpx>=0.27.2",
  "openpyxl>=3.1.5",
  "pyarrow>=20.0.0",
  "pydantic-settings>=2.6.0",
  "python-multipart>=0.0.20",
  "psycopg[binary]>=3.2.0",
  "redis>=5.2.0",
  "sse-starlette>=2.1.3",
  "sqlalchemy>=2.0.36",
  "temporalio>=1.9.0",
  "uvicorn[standard]>=0.32.0",
  "watchfiles>=1.1.1",
  "xlrd>=2.0.1",
]

[dependency-groups]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.7.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/nta_backend"]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
extend-exclude = ["migrations/versions"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
